Split implementation
